feat(tools): Add secure CLI admin tool
Implements admin authentication and refactors ICurrentUserService for console environment.
